Move `AgentStateField` to `stateful`
