+++ packages/engine/lib/stateful/src/agent/mod.rs
@@ -1,3 +1,4 @@
 mod schema;
+mod state;
 
-pub use self::schema::AgentSchema;
+pub use self::{schema::AgentSchema, state::AgentStateField};
