+++ packages/engine/src/datastore/arrow/batch_conversion.rs
@@ -19,7 +19,7 @@ use memory::arrow::{
 };
 use serde_json::value::Value;
 use stateful::{
-    agent::AgentSchema,
+    agent::{AgentSchema, AgentStateField},
     field::{FieldScope, FieldTypeVariant, RootFieldKey},
     message::MESSAGE_BATCH_SCHEMA,
 };
@@ -32,10 +32,7 @@ use crate::{
         schema::{EngineComponent, IsRequired},
         UUID_V4_LEN,
     },
-    hash_types::{
-        state::{AgentStateField, BUILTIN_FIELDS},
-        Agent,
-    },
+    hash_types::{state::BUILTIN_FIELDS, Agent},
     simulation::package::creator::PREVIOUS_INDEX_FIELD_KEY,
 };
