+++ packages/engine/src/datastore/schema/field_spec/built_in.rs
@@ -1,6 +1,6 @@
-use crate::hash_types::state::AgentStateField;
-
 // These are columns that are included in the schema even if the user opts out
+
+use stateful::agent::AgentStateField;
 
 pub const REQUIRED: [AgentStateField; 4] = [
     AgentStateField::AgentId,
